Qt 6.9.3 source code

+++ qtbase/src/widgets/itemviews/qtreewidget.cpp
@@ -42,7 +42,6 @@ public:
 
 /*!
     \enum QTreeWidgetItem::ChildIndicatorPolicy
-    \since 4.3
 
     \value ShowIndicator     The controls for expanding and collapsing will be shown for this item even if there are no children.
     \value DontShowIndicator   The controls for expanding and collapsing will never be shown even if there are children.  If the node is forced open the user will not be able to expand or collapse the item.
@@ -51,7 +50,6 @@ public:
 
 /*!
     \fn void QTreeWidgetItem::setDisabled(bool disabled)
-    \since 4.3
 
     Disables the item if \a disabled is true; otherwise enables the item.
 
@@ -60,7 +58,6 @@ public:
 
 /*!
     \fn bool QTreeWidgetItem::isDisabled() const
-    \since 4.3
 
     Returns \c true if the item is disabled; otherwise returns \c false.
 
@@ -742,6 +739,8 @@ Qt::DropActions QTreeModel::supportedDropActions() const
 
 void QTreeModel::itemChanged(QTreeWidgetItem *item)
 {
+    if (item->columnCount() <= 0)
+        return;
     SkipSorting skipSorting(this); //this is kind of wrong, but not doing this would kill performance
     QModelIndex left = index(item, 0);
     QModelIndex right = index(item, item->columnCount() - 1);
@@ -952,7 +951,6 @@ void QTreeModel::timerEvent(QTimerEvent *ev)
 
 /*!
     \fn void QTreeWidgetItem::sortChildren(int column, Qt::SortOrder order)
-    \since 4.2
 
     Sorts the children of the item using the given \a order,
     by the values in the given \a column.
@@ -969,7 +967,6 @@ void QTreeModel::timerEvent(QTimerEvent *ev)
 
 /*!
   \fn void QTreeWidgetItem::setSelected(bool select)
-  \since 4.2
 
   Sets the selected state of the item to \a select.
 
@@ -989,7 +986,6 @@ void QTreeWidgetItem::setSelected(bool select)
 
 /*!
   \fn bool QTreeWidgetItem::isSelected() const
-  \since 4.2
 
   Returns \c true if the item is selected, otherwise returns \c false.
 
@@ -1002,7 +998,6 @@ bool QTreeWidgetItem::isSelected() const
 
 /*!
   \fn void QTreeWidgetItem::setHidden(bool hide)
-  \since 4.2
 
   Hides the item if \a hide is true, otherwise shows the item.
   \note A call to this function has no effect if the item is not currently in a view. In particular,
@@ -1028,7 +1023,6 @@ void QTreeWidgetItem::setHidden(bool hide)
 
 /*!
   \fn bool QTreeWidgetItem::isHidden() const
-  \since 4.2
 
   Returns \c true if the item is hidden, otherwise returns \c false.
 
@@ -1050,7 +1044,6 @@ bool QTreeWidgetItem::isHidden() const
 
 /*!
   \fn void QTreeWidgetItem::setExpanded(bool expand)
-  \since 4.2
 
   Expands the item if \a expand is true, otherwise collapses the item.
   \warning The QTreeWidgetItem must be added to the QTreeWidget before calling this function.
@@ -1068,7 +1061,6 @@ void QTreeWidgetItem::setExpanded(bool expand)
 
 /*!
   \fn bool QTreeWidgetItem::isExpanded() const
-  \since 4.2
 
   Returns \c true if the item is expanded, otherwise returns \c false.
 
@@ -1085,7 +1077,6 @@ bool QTreeWidgetItem::isExpanded() const
 
 /*!
   \fn void QTreeWidgetItem::setFirstColumnSpanned(bool span)
-  \since 4.3
 
   Sets the first section to span all columns if \a span is true;
   otherwise all item sections are shown.
@@ -1103,7 +1094,6 @@ void QTreeWidgetItem::setFirstColumnSpanned(bool span)
 
 /*!
   \fn bool QTreeWidgetItem::isFirstColumnSpanned() const
-  \since 4.3
 
   Returns \c true if the item is spanning all the columns in a row; otherwise returns \c false.
 
@@ -1218,7 +1208,6 @@ bool QTreeWidgetItem::isFirstColumnSpanned() const
 
 /*!
     \fn QBrush QTreeWidgetItem::background(int column) const
-    \since 4.2
 
     Returns the brush used to render the background of the specified \a column.
 
@@ -1227,7 +1216,6 @@ bool QTreeWidgetItem::isFirstColumnSpanned() const
 
 /*!
     \fn void QTreeWidgetItem::setBackground(int column, const QBrush &brush)
-    \since 4.2
 
     Sets the background brush of the label in the given \a column to the
     specified \a brush.
@@ -1243,7 +1231,6 @@ bool QTreeWidgetItem::isFirstColumnSpanned() const
 
 /*!
     \fn QBrush QTreeWidgetItem::foreground(int column) const
-    \since 4.2
 
     Returns the brush used to render the foreground (e.g. text) of the
     specified \a column.
@@ -1255,7 +1242,6 @@ bool QTreeWidgetItem::isFirstColumnSpanned() const
 
 /*!
     \fn void QTreeWidgetItem::setForeground(int column, const QBrush &brush)
-    \since 4.2
 
     Sets the foreground brush of the label in the given \a column to the
     specified \a brush.
@@ -1281,7 +1267,6 @@ bool QTreeWidgetItem::isFirstColumnSpanned() const
 
 /*!
   \fn QSize QTreeWidgetItem::sizeHint(int column) const
-  \since 4.1
 
   Returns the size hint set for the tree item in the given
   \a column (see \l{QSize}).
@@ -1289,7 +1274,6 @@ bool QTreeWidgetItem::isFirstColumnSpanned() const
 
 /*!
   \fn void QTreeWidgetItem::setSizeHint(int column, const QSize &size)
-  \since 4.1
 
   Sets the size hint for the tree item in the given \a column to be \a size.
   If no size hint is set or \a size is invalid, the item
@@ -1895,8 +1879,6 @@ void QTreeWidgetItem::write(QDataStream &out) const
 #endif // QT_NO_DATASTREAM
 
 /*!
-    \since 4.1
-
     Constructs a copy of \a other. Note that type() and treeWidget()
     are not copied.
 
@@ -2032,8 +2014,6 @@ QTreeWidgetItem *QTreeWidgetItem::takeChild(int index)
 }
 
 /*!
-  \since 4.1
-
   Appends the given list of \a children to the item.
 
   \sa insertChildren(), takeChildren()
@@ -2044,8 +2024,6 @@ void QTreeWidgetItem::addChildren(const QList<QTreeWidgetItem*> &children)
 }
 
 /*!
-  \since 4.1
-
   Inserts the given list of \a children into the list of the item children at \a index .
 
   Children that have already been inserted somewhere else won't be inserted.
@@ -2099,8 +2077,6 @@ void QTreeWidgetItem::insertChildren(int index, const QList<QTreeWidgetItem*> &c
 }
 
 /*!
-  \since 4.1
-
   Removes the list of children and returns it, otherwise returns an empty list.
 */
 QList<QTreeWidgetItem*> QTreeWidgetItem::takeChildren()
@@ -2221,8 +2197,6 @@ QVariant QTreeWidgetItem::childrenCheckState(int column) const
 }
 
 /*!
-  \since 4.5
-
   Causes the model associated with this item to emit a
   \l{QAbstractItemModel::dataChanged()}{dataChanged}() signal for this
   item.
@@ -2562,8 +2536,6 @@ void QTreeWidgetPrivate::dataChanged(const QModelIndex &topLeft,
 */
 
 /*!
-  \since 4.3
-
   \fn void QTreeWidget::removeItemWidget(QTreeWidgetItem *item, int column)
 
   Removes the widget set in the given \a item in the given \a column.
@@ -2643,8 +2615,6 @@ void QTreeWidget::setColumnCount(int columns)
 }
 
 /*!
-  \since 4.2
-
   Returns the tree widget's invisible root item.
 
   The invisible root item provides access to the tree widget's top-level items
@@ -2702,8 +2672,6 @@ void QTreeWidget::insertTopLevelItem(int index, QTreeWidgetItem *item)
 }
 
 /*!
-    \since 4.1
-
     Appends the \a item as a top-level item in the widget.
 
     \sa insertTopLevelItem()
@@ -2740,8 +2708,6 @@ int QTreeWidget::indexOfTopLevelItem(QTreeWidgetItem *item) const
 }
 
 /*!
-  \since 4.1
-
   Inserts the list of \a items at \a index in the top level in the view.
 
   Items that have already been inserted somewhere else won't be inserted.
@@ -2827,7 +2793,6 @@ void QTreeWidget::setHeaderLabels(const QStringList &labels)
 
 /*!
     \fn void QTreeWidget::setHeaderLabel(const QString &label)
-    \since 4.2
 
     Same as setHeaderLabels(QStringList(\a label)).
 */
@@ -2844,7 +2809,6 @@ QTreeWidgetItem *QTreeWidget::currentItem() const
 }
 
 /*!
-    \since 4.1
     Returns the current column in the tree widget.
 
     \sa setCurrentItem(), columnCount()
@@ -2868,7 +2832,6 @@ void QTreeWidget::setCurrentItem(QTreeWidgetItem *item)
 }
 
 /*!
-  \since 4.1
   Sets the current \a item in the tree widget and the current column to \a column.
 
   \sa currentItem()
@@ -2880,7 +2843,6 @@ void QTreeWidget::setCurrentItem(QTreeWidgetItem *item, int column)
 }
 
 /*!
-  \since 4.4
   Sets the current \a item in the tree widget and the current column to \a column,
   using the given \a command.
 
@@ -2933,8 +2895,6 @@ QRect QTreeWidget::visualItemRect(const QTreeWidgetItem *item) const
 }
 
 /*!
-  \since 4.1
-
   Returns the column used to sort the contents of the widget.
 
   \sa sortItems()
@@ -3014,8 +2974,6 @@ bool QTreeWidget::isPersistentEditorOpen(QTreeWidgetItem *item, int column) cons
 }
 
 /*!
-    \since 4.1
-
     Returns the widget displayed in the cell specified by \a item and the given \a column.
 
     \sa setItemWidget(), removeItemWidget()
@@ -3027,8 +2985,6 @@ QWidget *QTreeWidget::itemWidget(QTreeWidgetItem *item, int column) const
 }
 
 /*!
-    \since 4.1
-
     Sets the given \a widget to be displayed in the cell specified by the given
     \a item and \a column.
 
@@ -3094,8 +3050,6 @@ QList<QTreeWidgetItem*> QTreeWidget::findItems(const QString &text, Qt::MatchFla
 
 
 /*!
-  \since 4.3
-
   Returns the item above the given \a item.
 */
 QTreeWidgetItem *QTreeWidget::itemAbove(const QTreeWidgetItem *item) const
@@ -3109,8 +3063,6 @@ QTreeWidgetItem *QTreeWidget::itemAbove(const QTreeWidgetItem *item) const
 }
 
 /*!
-  \since 4.3
-
   Returns the item visually below the given \a item.
 */
 QTreeWidgetItem *QTreeWidget::itemBelow(const QTreeWidgetItem *item) const
